feat: Store Bun Alerts to and Get Bun Alerts from Firestore (#7)
* fix: Elevate Location Services Checking to App.tsx

Previously there were weird issues where there would never be a response from the navigation.geolocation.getCurrentLocation() function being used to check if permissions have been granted, causing an error message to display when changing pages at times. By elevating this check to the root of the webapp. the goal is to check only once on load and avoid needing to check on each page load to reduce the likelihood of errors impacting the service.

* fix: Convert rankOfSmallestBun to Enum

This will make it easier to order buns by rank and run unique Baby Bun logic in the future.

* feat: Store Bun Alerts to and Get Bun Alerts from Firestore